Split worldbuilding out of the IPC monolith
The largest extraction so far: 193 handlers across thirteen channel prefixes that
together are one vault type — characters, groups, maps, world, worldChat,
encyclopedia, story, threads, scenes, manuscript, rules, continuity, questions.

  electron/ipc.ts     1333-1892 -> electron/ipc/worldbuilding.ts     (193 handlers)
  electron/preload.ts   202-430 -> electron/preload/worldbuilding.ts (194 bindings)
  shared/types.ts     7388-7763 -> shared/api/worldbuilding.ts       (193 signatures)

Taken prefix by prefix, several of these looked hopelessly interleaved — characters:
alone reported 159 handlers of other domains inside its range. Taken together they
are one contiguous, pure block, because what surrounded each was the rest of
worldbuilding. That is the general lesson for what remains: the interleaving is
domains taking turns, not disorder, so the unit of extraction is the product surface,
not the channel prefix.

One prefix genuinely is shared and had to be split by full channel name: places: serves
both the genealogical dossier (places:listForPerson, places:add) and invented worlds
(places:listWorld, places:inhabitants). Only the six world channels moved.

activeWorldChatRequestId moved from preload.ts into the slice, its only user, like the
databases equivalent before it.

Six more tests read the hot files as text, three of them through a local `read`
helper rather than an inline path; the helper now routes through readSource.

Where the split stands: 594 of 1,206 channels — just under half — live in a domain
module, 612 remain in electron/ipc.ts. That file is 2,841 lines (from 6,013 — down
53%), electron/preload.ts 1,068 (from 1,862 — down 43%), shared/types.ts 8,366 (from
9,553).

Verified: 1,425 tests, lint and typecheck clean, contract test green.
