+++ scripts/test-world-chat-parity.mjs
@@ -3,16 +3,19 @@ import test from 'node:test';
 import { readFile } from 'node:fs/promises';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
+import { readSource } from './ipc-channel-census.mjs';
 
 const root = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
-const read = (file) => readFile(path.join(root, file), 'utf8');
+// readSource resolves the '@main' / '@bridge' / '@api' sentinels to whole surfaces —
+// the three former hot files are directories now — and any other path to that file.
+const read = async (file) => readSource(file);
 
 test('world chat has the same conversation, model, context and streaming controls as study chat', async () => {
   const [view, types, preload, ipc] = await Promise.all([
     read('src/views/WorldChatView.tsx'),
-    read('shared/types.ts'),
-    read('electron/preload.ts'),
-    read('electron/ipc.ts'),
+    read('@api'),
+    read('@bridge'),
+    read('@main'),
   ]);
   for (const contract of [
     'listWorldChatConversations',
